add /api/posts helper
